+++ app/Http/Controllers/Events/EventICalExport.php
@@ -6,6 +6,8 @@ namespace App\Http\Controllers\Events;
 
 use App\Http\Controllers\Controller;
 use App\Livewire\Profile\CalendarLinks;
+use App\Models\Configuration;
+use App\Models\ConfigurationKey;
 use App\Models\Filter\MemberFilter;
 use App\Models\Member;
 use App\Models\User;
@@ -26,7 +28,8 @@ class EventICalExport extends Controller
 
         $authToken = $this->getPersonalAccessToken();
 
-        if ($authToken && $this->hasMemberShowPermission($authToken)) {
+        if (Configuration::getBool(ConfigurationKey::EVENT_BIRTHDAYS_IN_ICS_EXPORT) ||
+            (($authToken = $this->getPersonalAccessToken()) && $this->hasMemberShowPermission($authToken))) {
             $this->addMembersToCalendar($calendar);
         }
 
@@ -77,7 +80,7 @@ class EventICalExport extends Controller
             //last year birthday
             $calEvent = Event::create($member->getFullName());
             $calEvent->startsAt($birthdayLastYear);
-            $calEvent->description((string) ($currentYear - 1 - $member->birthday->year));
+            $calEvent->description((string)($currentYear - 1 - $member->birthday->year));
             $calEvent->fullDay();
             $calendar->event($calEvent);
 
@@ -86,7 +89,7 @@ class EventICalExport extends Controller
             $birthday->setYear($currentYear);
             $calEvent = Event::create($member->getFullName());
             $calEvent->startsAt($birthday);
-            $calEvent->description((string) ($currentYear - $member->birthday->year));
+            $calEvent->description((string)($currentYear - $member->birthday->year));
             $calEvent->fullDay();
             $calendar->event($calEvent);
 
@@ -95,7 +98,7 @@ class EventICalExport extends Controller
             $birthdayNextYear->setYear($currentYear + 1);
             $calEvent = Event::create($member->getFullName());
             $calEvent->startsAt($birthdayNextYear);
-            $calEvent->description((string) ($currentYear + 1 - $member->birthday->year));
+            $calEvent->description((string)($currentYear + 1 - $member->birthday->year));
             $calEvent->fullDay();
             $calendar->event($calEvent);
         }
@@ -136,7 +139,7 @@ class EventICalExport extends Controller
     private function getEventList($inclLoggedInOnly = false): Collection
     {
         $eventList = \App\Models\Event::query()->orderBy('start', 'desc');
-        if (! $inclLoggedInOnly) {
+        if (!$inclLoggedInOnly) {
             $eventList = $eventList->where('logged_in_only', false);
         }
         $eventList = $eventList->where('enabled', true);