+++ app/Models/ConfigurationKey.php
@@ -10,6 +10,7 @@ enum ConfigurationKey: string
     case EVENT_FILTER_DEFAULT_START_TODAY = "eventFilter_defaultStartToday";
     case EVENT_FILTER_DEFAULT_START_DATE = "eventFilter_defaultStartDate";
     case EVENT_FILTER_DEFAULT_END_DATE = "eventFilter_defaultEndDate";
+    case EVENT_BIRTHDAYS_IN_ICS_EXPORT = "eventFilter_birthdaysInIcsExport";
     case POLL_PUBLIC_FILTER_BEFORE_ENTRANCE = "publicPoll_beforeEntrance";
     case POLL_PUBLIC_FILTER_AFTER_RETIRED = "publicPoll_afterRetired";
     case POLL_PUBLIC_FILTER_SHOW_PAUSED = "publicPoll_showPaused";