+++ resources/views/livewire/settings/events.blade.php
@@ -5,7 +5,7 @@
         </h2>
 
         <p class="mt-1 text-sm text-gray-600">
-            {{ __('Show and edit user global event settings.') }}
+            {{ __('Show and edit global event settings.') }}
         </p>
     </header>
 
@@ -34,6 +34,16 @@
         <div class="flex items-center justify-between border-t border-slate-400/20 py-3">
             <span>{{__("End date")}}</span>
             <x-input-date wire:model.live="eventEndDate"/>
+        </div>
+        <div class="flex items-center justify-between border-t border-slate-400/20 py-3"
+             x-init="enabled={{$birthdaysInPublicICS ? 'true':'false'}}"
+             x-data="{enabled:false,
+                                    switchChanged(curState) {
+                                        this.enabled = curState;
+                                        $wire.setBirthdayInIcsExport(curState);
+                                    }}">
+            <span>{{__("Show members birthdays in public ics export")}}</span>
+            <x-input-switch/>
         </div>
     </div>
 </section>